# Design/middleBackendSwagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MiniMLBackend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ML/ranges/{simId}": {
      "post": {
        "tags": [
          "ML"
        ],
        "parameters": [
          {
            "name": "simId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MLDateSpecifications"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MLDateSpecifications"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MLDateSpecifications"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ML/train/{simId}": {
      "post": {
        "tags": [
          "ML"
        ],
        "parameters": [
          {
            "name": "simId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MLDateSpecifications"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MLDateSpecifications"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MLDateSpecifications"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ML/validate/{simId}": {
      "post": {
        "tags": [
          "ML"
        ],
        "parameters": [
          {
            "name": "simId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ML/model/{simId}": {
      "get": {
        "tags": [
          "ML"
        ],
        "parameters": [
          {
            "name": "simId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ML/validationresult/{simId}": {
      "post": {
        "tags": [
          "ML"
        ],
        "parameters": [
          {
            "name": "simId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ML/predictstream/{simId}": {
      "post": {
        "tags": [
          "ML"
        ],
        "parameters": [
          {
            "name": "simId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MLDateSpecifications": {
        "type": "object",
        "properties": {
          "trainStart": {
            "type": "string",
            "format": "date-time"
          },
          "trainEnd": {
            "type": "string",
            "format": "date-time"
          },
          "testStart": {
            "type": "string",
            "format": "date-time"
          },
          "testEnd": {
            "type": "string",
            "format": "date-time"
          },
          "validStart": {
            "type": "string",
            "format": "date-time"
          },
          "validEnd": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}